GafferTest : Added CountingExecutableNode class.
This is split out from ExecutableNodeTest and will be used to help write other test cases.
