+++ python/GafferTest/ExecutableNodeTest.py
@@ -44,48 +44,10 @@ import GafferTest
 
 class ExecutableNodeTest( GafferTest.TestCase ) :
 
-	class MyNode( Gaffer.ExecutableNode ) :
-
-		def __init__( self, withHash, requiresSequenceExecution = False ) :
-
-			Gaffer.ExecutableNode.__init__( self )
-
-			self.__requiresSequenceExecution = requiresSequenceExecution
-
-			self.__withHash = withHash
-			self.executionCount = 0
-
-		def execute( self ) :
-
-			self.executionCount += 1
-
-		def executeSequence( self, frames ) :
-
-			if not self.__requiresSequenceExecution :
-				Gaffer.ExecutableNode.executeSequence( self, frames )
-				return
-
-			self.executionCount += 1
-
-		def hash( self, context ) :
-
-			if not self.__withHash :
-				return IECore.MurmurHash()
-
-			h = Gaffer.ExecutableNode.hash( self, context )
-			h.append( context.getFrame() )
-			return h
-
-		def requiresSequenceExecution( self ) :
-
-			return self.__requiresSequenceExecution
-
-	IECore.registerRunTimeTyped( MyNode )
-
 	def testIsExecutable( self ) :
 
-		self.assertTrue( issubclass( self.MyNode, Gaffer.ExecutableNode ) )
-		self.assertTrue( isinstance( self.MyNode( True ), Gaffer.ExecutableNode ) )
+		self.assertTrue( issubclass( GafferTest.CountingExecutableNode, Gaffer.ExecutableNode ) )
+		self.assertTrue( isinstance( GafferTest.CountingExecutableNode(), Gaffer.ExecutableNode ) )
 
 	def testHash( self ) :
 
@@ -97,31 +59,31 @@ class ExecutableNodeTest( GafferTest.TestCase ) :
 		c3.setFrame( 3.0 )
 
 		# hashes that don't use the context are equivalent
-		n = ExecutableNodeTest.MyNode(False)
+		n = GafferTest.CountingExecutableNode( withHash = False )
 		self.assertEqual( n.hash( c1 ), n.hash( c1 ) )
 		self.assertEqual( n.hash( c1 ), n.hash( c2 ) )
 		self.assertEqual( n.hash( c1 ), n.hash( c3 ) )
 
 		# hashes that do use the context differ
-		n2 = ExecutableNodeTest.MyNode(True)
+		n2 = GafferTest.CountingExecutableNode( withHash = True )
 		self.assertEqual( n2.hash( c1 ), n2.hash( c1 ) )
 		self.assertNotEqual( n2.hash( c1 ), n2.hash( c2 ) )
 		self.assertNotEqual( n2.hash( c1 ), n2.hash( c3 ) )
 
 		# hashes match across the same node type
-		n3 = ExecutableNodeTest.MyNode(True)
+		n3 = GafferTest.CountingExecutableNode( withHash = True )
 		self.assertEqual( n2.hash( c1 ), n3.hash( c1 ) )
 		self.assertEqual( n2.hash( c2 ), n3.hash( c2 ) )
 		self.assertEqual( n2.hash( c3 ), n3.hash( c3 ) )
 
 		# hashes differ across different node types
-		class MyNode2( ExecutableNodeTest.MyNode ) :
+		class MyNode( GafferTest.CountingExecutableNode ) :
 			def __init__( self ) :
-				ExecutableNodeTest.MyNode.__init__( self, True )
+				GafferTest.CountingExecutableNode.__init__( self )
 
-		IECore.registerRunTimeTyped( MyNode2 )
+		IECore.registerRunTimeTyped( MyNode )
 
-		n4 = MyNode2()
+		n4 = MyNode()
 
 		self.assertNotEqual( n4.hash( c1 ), n3.hash( c1 ) )
 		self.assertNotEqual( n4.hash( c2 ), n3.hash( c2 ) )
@@ -129,7 +91,7 @@ class ExecutableNodeTest( GafferTest.TestCase ) :
 
 	def testExecute( self ) :
 
-		n = ExecutableNodeTest.MyNode(True)
+		n = GafferTest.CountingExecutableNode()
 		self.assertEqual( n.executionCount, 0 )
 
 		n.execute()
@@ -146,7 +108,7 @@ class ExecutableNodeTest( GafferTest.TestCase ) :
 
 	def testExecuteSequence( self ) :
 
-		n = ExecutableNodeTest.MyNode(True)
+		n = GafferTest.CountingExecutableNode()
 		self.assertEqual( n.executionCount, 0 )
 
 		n.executeSequence( [ 1, 2, 3 ] )
@@ -156,7 +118,7 @@ class ExecutableNodeTest( GafferTest.TestCase ) :
 		self.assertEqual( n.executionCount, 6 )
 
 		# requiring execution doesn't tally the count per frame
-		n2 = ExecutableNodeTest.MyNode( True, requiresSequenceExecution = True )
+		n2 = GafferTest.CountingExecutableNode( requiresSequenceExecution = True )
 		self.assertEqual( n2.executionCount, 0 )
 
 		n2.executeSequence( [ 1, 2, 3 ] )
@@ -167,7 +129,7 @@ class ExecutableNodeTest( GafferTest.TestCase ) :
 
 	def testRequiresSequenceExecution( self ) :
 
-		n = ExecutableNodeTest.MyNode(True)
+		n = GafferTest.CountingExecutableNode()
 		self.assertEqual( n.requiresSequenceExecution(), False )
 
 	def testRequirements( self ) :
@@ -178,8 +140,8 @@ class ExecutableNodeTest( GafferTest.TestCase ) :
 		c2 = Gaffer.Context()
 		c2.setFrame( 2 )
 
-		n = ExecutableNodeTest.MyNode(True)
-		n2 = ExecutableNodeTest.MyNode(True)
+		n = GafferTest.CountingExecutableNode()
+		n2 = GafferTest.CountingExecutableNode()
 
 		# make n2 require n
 		n2["requirements"][0].setInput( n['requirement'] )
@@ -254,7 +216,7 @@ class ExecutableNodeTest( GafferTest.TestCase ) :
 		self.assertTrue( t5 in s )
 
 		# MyNode.hash() depends on the context time, so tasks will vary
-		my = ExecutableNodeTest.MyNode( True )
+		my = GafferTest.CountingExecutableNode()
 		c.setFrame( 1 )
 		t1 = Gaffer.ExecutableNode.Task( my, c )
 		t2 = Gaffer.ExecutableNode.Task( my, c )
@@ -263,7 +225,7 @@ class ExecutableNodeTest( GafferTest.TestCase ) :
 		c2.setFrame( 2 )
 		t3 = Gaffer.ExecutableNode.Task( my, c2 )
 		self.assertNotEqual( t1, t3 )
-		my2 = ExecutableNodeTest.MyNode( True )
+		my2 = GafferTest.CountingExecutableNode()
 		t4 = Gaffer.ExecutableNode.Task( my2, c2 )
 		self.assertNotEqual( t1, t4 )
 		self.assertEqual( t3, t4 )
